Fix ObjectTracker to allow retry on error
When Watch is called, the object's GroupKind is stored in a cache, so that duplicate
registration is avoided. But if some reason the watch registration errored out, the
cache doesn't clear out the object as the wrong key was passed. This fix allows the
Watch to be retried correctly in the case of failures with the registration.