⚠️ Update klog dependency to v2 (#4284)
* Switch to klog v2

This switches klog usage over to klog v2.5.0 to match the current
version being used in k/k.

This should also address some issues we're seeing that have since been
fixed in klog between the quite old v1.0.0 and v2.5.0

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>

* Update test/infra/docker go.mod for klog v2

Missed updating the go.mod file in test/infrastructure/docker/go.mod to
match upgrading the rest of the repo to use klog v2.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>

* Updates to docker go.sum

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>

* Not klog v2 change in v1alpha4 provider docs

This updates the v1alpha3-to-v1alpha4 docs to call out the move to
klogv2 and recommending that all cluster providers do the same.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>