General docs cleanup
Reading through the book for the first time, I took the opportunity to
look at it with fresh eyes and find things that were missing or needed
an update.

This addresses various grammer, spelling, and linking issues found
throughout the book. There are some mdlint updates that were included
where they were very obvious, but that was not the scope of this effort.
If we want to go with full mdlinting, then that should be its own PR and
also include a job to make sure it's enforced.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>