Make conversion-gen output location explicit
The conversion-gen command has an --output-base argument to control
where generated files are placed. The default value for this argument
can vary depending on whether or not $GOPATH is set or not. This results
in our generated files being created in the wrong place for some people
in a subtle and non-obvious way.

For those running `make generate` with GOPATH set, the files are created
in `$GOPATH/src/[file path]`. For those without GOPATH set, files are
created where we expect them to be within the repo at `./[file path]`.

To make sure files are always generated to the location where we expect
them to be, this updates our calls to conversion-gen to always be
explicit and set `--output-base=[path to code]`. This actually causes
issues if the code is actually located under GOPATH, so only add
`--output-path` if we see we are not under GOPATH.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>