Update tilt development instructions
The current tilt development instructions point to a missing section in
the Quick Start guide. The quick start instructions use clusterctl to
create worker cluster configs, which is not supported.

This updates the instructions to point out the example configs that are
kept in-repo and warns not to use clusterctl.

Also bumps our example config files to use the v1alpha4 apiVersion.

Raises tilt min version to v0.16.0

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>