Add helper script to tilt up CAPD environment
There are several steps needed when using tilt for working with CAPD
that are not always obvious. Some docs are spread out between the tilt
docs and the getting started guide, and some of the minor details are
easy to overlook.

This adds a helper script to make it easy to start up a CAPD tilt
run from a fresh environment.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>