Use docker API instead of CLI in test framework
We have several places in the test framework where we exec out to the
docker CLI to perform various operations. This removes the dependency on
that CLI being present by updating these instances to use the Docker
Engine SDK to perform equivalent operations.

This moves the code that interacts with the Docker engine into its own
module. This will be where additional work will be done to provide a
generic interface to allow better unit testing and more container
runtime support that can be used by both the test framework as well as
CAPD.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>