:seedling: Add a fuzzer function for *metav1.Time
Currently, when metav1.Time is a pointer it never gets fuzzed by the
meta functions. After a further look, it seems like the Creation or
Deletion timestamps on ObjectMeta are fuzzed in this way as well.

This patch adds custom logic to fuzz the metav1.Time objects when it's
not nil.

Signed-off-by: Vince Prignano <vincepri@vmware.com>