test/e2e: pin cgroupDriver to cgroupfs
Starting with kind v0.11, kind images will have cgroup-root=/kubelet
hard-coded to fix an issue with nested cgroup hierarchies. The hierarchy
below /kubelet is not compatible with systemd cgroupfs (in contrast to
/).
Starting with kubeadm v1.21, kubeadm will default to cgroupDriver
systemd.

Thus, the combination of images from kind >=v0.11 and kubeadm >=v1.21 are
not compatible. To solve this problem we will pin cgroupDriver to cgroupfs
until kind supports the systemd hierarchy with the /kubelet cgroup-root.