Enable nilerr lint checker
The nilerr linter checks whether there is a check if an error is not
nil, but then does not return the error. This is legitimate in some
cases, but can be an error in others. Enabling this check forces us to
be explicit about our intent if we do mean to return without returning
the error.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>