⚠️ Refactor golangci-lint config to remove false negatives
During some audit of golangci-lint, it was found that one of the exclude
directives was matching a lot more than it should have.

The removal of the unwanted regex caused a large backlog of linting
errors to show up. This changeset brings in a refactor in the
golangci-lint configuration to make sure we catch more errors in the
linting process and enable more linters over time.

Signed-off-by: Vince Prignano <vincepri@vmware.com>