:bug: Use unstructured for conversion utils tests, and fix how we set annotations
During a rewrite of the tests for the conversion util codebase which
sets and resets the annotation to retain data when doing conversion,
opted to use unstructured objects with a generic old "Machine" type
instead of importing multiple API packages.

The switch to unstructured showed that we weren't setting annotation
properly by calling SetAnnotations, but rather we opted to change the
map coming from GetAnnotations() directly in some cases.

Signed-off-by: Vince Prignano <vincepri@vmware.com>