:seedling: Use client-go global scheme as much as possible
A test is currently flaking and reporting a race condition because we
use the scheme in a way that shouldn't be used. A runtime.Scheme object
is not safe to use concurrently (add and read at the same time), but
it's meant to be used in an immutable way after the first registration
happens. In some tests we were reusing and adding to the same scheme
while other tests were reading from it, this caused a race condition to
trigger given that the scheme uses maps internally to store its
information.

Going forward, we should standardize on using the global scheme more and
more to keep things simple. This would not only simplify our tests, but
also simplify our code.

Signed-off-by: Vince Prignano <vincepri@vmware.com>