Machine status.phase should be provisioned when there is a ProviderID
We're currently in a situation where the "provisioned" phase would never
show up given the sequence of events between the controllers. The
proposal expects "provisioned" to be set whenever a Machine has
a spec.providerID field populated, the infrastructure provider is
reporting ready, and there is no Node associated with the Machine.

Signed-off-by: Vince Prignano <vincepri@vmware.com>