Use docker API instead of exec'ing docker cmds
This gets rid of calls out to the docker CLI with parsing of its
output to perform docker management. Calls now use the docker SDK
directly.

This is the first step towards making the CAPD more testable and
potentially able to use other CRIs. Ideally, long term, we will want to
refactor the code to move all runtime interaction into one or more
interfaces that would give us the option for swapping out
implementations.

We can also then set up better unit testing for all code paths to make
sure the interacting code is correct, regardless of the implementation
of the runtime interaction.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>