CAPD: set container restartPolicy to unless-stopped
If the docker engine is restarted, either through a service restart or a
full system reboot, the containers created by CAPD are left in a stopped
state.

This adds the restartPolicy setting to the Docker container creation so
that containers will be automatically restarted on service restart
unless the user has explicitly stopped the container prior to the
restart.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>