:seedling: KubeadmControlPlane internal/proxy should use pointer structs
The net.Conn interface doesn't need a plain struct, it can work well
with pointers as well. To avoid copying structs around every time we
initiate a connection or call a method, convert the underlying Conn
struct in the proxy package to using pointer receiving methods.

This change also fixes an issue with read/writeDeadline, which was being
set in the local copy of the Conn struct before instead of the current
object, which created an ineffective assignment.

Signed-off-by: Vince Prignano <vincepri@vmware.com>