Added comments to `WatchFilterValue` field.
What this PR does / why we need it:
This PR adds comments to all the WatchFilterValue in all the Reconciler types. This was needed to be done for a better explanation of its purpose.

Which issue(s) this PR fixes (optional, in fixes #<issue number>(, fixes #<issue_number>, ...) format, will close the issue(s) when PR gets merged):
Fixes #5112