:seedling: Add conversion-verifier to always verify conversion code
This change adds a sort-of-custom linter to the hack/tools folder which
does basic verification of our types. The linter isn't generic and it's
customized to the Cluster API's codebase needs and structure. It can be
improved and made better in the future and we should be able to further
generalize it with time.

The conversion-verifier is needed given that conversion information has
been missed in the past causing types to remain unconvertible.

Signed-off-by: Vince Prignano <vincepri@vmware.com>