Templating for CRS rendered objects.
This adds templating to CRS objects as they are being inserted into the
remote cluster.

Standard Go templating is used, and cluster is made available as the
context, along with a couple of helper functions to help with getting
annotations and labels from the cluster.

This can be useful if you want to create objects in the remote cluster
that are dynamic, based on the cluster that has been created.