Add test coverage for container runtime calls
This builds on the CAPD changes to use the docker API instead of calling
out to the CLI. With those changes, it is now possible to insert a
testing fake and perform unit testing for the code that would normally
call out to Docker.

The test coverage introduced by this change is not extensive, but it
does add coverage where we previously had none at all. With this basic
coverage in place, it should now be easier to add new tests if we find
specific scenarios that we want to reproduce and validate for any new
bug reports.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>