Enforce godoc comments for exported functions
This updates a filter in our golangci-lint configuration to start
enforcing the presence of godoc comments on any functions or variables
that are exported. This aligns with most of the rest of the codebase and
makes it so there is (somewhat) useful documentation for anything that
can be accessed externally from a package.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>