Set gci linter prefix value
This sets the correct prefix value for the gci linter's namespace.

Good practice is to have imports separated into three distinct groups:
standard library imports, third party package imports, and local
imports. The gci linter is used to enforce this convention, making it
easy to see what is local, what is standard with go, and what is being
pulled in externally.

The gci linter configuration has a "prefix" setting that is used to tell
it what to match as the starting string to identify local imports. This
was set to "false", effectively disabling the ability to differentiate
local imports from third party ones.

This change sets the "prefix" value to our "sigs.k8s.io/cluster-api"
path so that our local imports are identified and grouped separately.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>