Enable more gocritic lint checks
There are a multiple groups of checks run by gocritic. Some are enabled
by default, but others require explicit enabling. This enables the
experimental group, minus a few exceptions that would either be too
disruptive or are things we aren't concerned about.

One of the main benefits of this extra coverage is the flagging of using
the deprecated ioutil package.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>