🐛 Requeue after 20 seconds if ControlPlane is not initialized yet
There was a fix https://github.com/kubernetes-sigs/cluster-api/pull/4947 a while ago, which has removed `Requeue: true`
from the return result in the machine controller.

While this change makes sense, it actually makes this reconcilation
loop pretty fragile: if the machines are not updated after `initialized`
is set in the control plane provier, the machine controller never starts
workload tracker and never updates the machine to be ready.

We have this issue when TalosControlPlane contoller runs along with
Sidero infrastructure provider.

Signed-off-by: Artem Chernyshev <artem.chernyshev@talos-systems.com>