fix: MachineDeploymentSpec.Replicas had an always-set default value
removes the default value of `machinedeployment.spec.replicas` as it breaks e.g. a gitops engine patcheing a MachineDeployment the cluster-autoscaler already reconfigured, resulting in a cluster scaling way down and then back-up again.
The gitops (in our case flux2) reconciliation returns to the default value of 1 when the field is not set in the manifest, cluster-autoscaler fixes this in its next run -> almost all cluster nodes get replace.

Signed-off-by: Matthias Riegler <matthias.riegler@traefik.io>