update github.com/go-logr/logr and k8s.io packages
Update github.com/go-logr/logr to the latest release, v1.2.2, and k8s.io
packages to the versions required to work with the v1.x versions of
github.com/go-logr/logr. The changes to accommodate
github.com/go-logr/logr were cherry picked from commit
3b4953a9021de8abf184b5fba59ff43e36738667.

In addition to the changes from commit
3b4953a9021de8abf184b5fba59ff43e36738667, this also required two small
changes to do a deep copy of a machine before calling
reconcileInterruptibleNodeLabel and changes to use the expected types
for webhook configurations required to work with the updated
controller-runtime.