Consistent ordering for deletion priority
If two machines had the same deletion priority, then on subsequent syncReplicas
calls different machines could be deleted. For example, when using Oldest or
Newest deletion priorities, then machines that are already deleting and machines
without a node ref have the same priority. Then this could lead to multiple
machines getting deleted on different calls to syncReplicas.

That is, if a MachineSet is scaled down by 1, then one machine would get
deleted, but then a second call to syncReplicas could delete a machine with no
node ref instead of the machine that is already deleting.

Signed-off-by: Donnie Adams <donald.g.adams@me.com>