:seedling: update talos/sidero references
Talos Systems is now Sidero Labs and we just renamed our github org.
This PR updates all of the references to make sure they align with the
new org name. Saw an issue where clusterctl couldn't follow the
redirects to the yaml assets, which appears to also occur when manually
curling (assuming a github bug).

Signed-off-by: Spencer Smith <spencer.smith@talos-systems.com>