:bug: Fix cluster reconcilation predicates
The current implementation for `ClusterUpdateUnpaused` is filtering all
cluster updates except for the case when `spec.paused` is updated from
`true` to `false`.

As all `Cluster` updates do not trigger `Machines` reconcilation,
setting `ControlPlaneInitialized` to `True` does not start workload nodes
watch in `MachinesController`.

That leads to cluster deployment being stuck and hanging until any other
unrelated event triggers that reconcilation.

Introduce a new predicate that triggers reconcilation when
`ControlPlaneInitialized` condition is set on a cluster.

Signed-off-by: Artem Chernyshev <artem.chernyshev@talos-systems.com>