Don't retry health check when Unauthorized is returned
This fixes unnecessary delays in self-repair when the underlying provider
(e.g. the AWS provider) performs periodic kubeconfig refreshes. Without
this, the health check is retried 10 times with a 10s interval, meaning
that the controller is unable to act for at least 100 seconds, even though
it could repair itself immediately.