add explicit securitycontexts to controllers
Add explicit, secure securityContexts for all managers except CAPD,
which is privileged and for testing purposes. These securityContexts
do not change the configuration, just make it explicit and enforced,
except for the seccompPolicy which changes from Unconfined to
RuntimeDefault. Syscalls filtered by RuntimeDefault policy are 95%
namespaced and require capabilities (which we drop) in the first place,
so no practical change there either.