E2e: Unhold CI packages before installing
It is common to mark kubeadm and related packages as held at a specific
version to avoid accidental upgrades. The script for injecting CI
artifacts does not take this into account, which means that if the base
image used already has some version of a package installed and held, the
script will fail to install the asked for version.
To avoid this, we unhold before installing the packages and then hold
again, as customary. Holding after installation also avoids any
unintended upgrades later.