Add a check that replicas are set in the MachinePool spec
Now if for some reason, replicas are not set, controller starts to
panic, and the pod goes into CrashLoopBackoff mode.

To prevent this, we add a check to verify that replicas value is not
nil. It allows to write a log message, return an error, and start
reconcilation again instead of panicing.