test/e2e check for machines being ready after provisioning on Runtime SDK test
The topology controller checks if the ControlPlane IsScaling() or any MachineDeployment
IsRollingOut() before running the BeforeClusterUpgrade hook. In some cases in e2e tests
the nodes were not ready yet and the timeout for the BeforeClusterUpgrade hook was not
sufficient for the nodes to get ready and running the hook.