Make release notes generation concurrent
Now that we retrieve the area labels from the PRs, the release notes
generation can take up to 2 minutes for big releases like 1.5. Although
in general this is not a problem since it should be a one time process,
it does slow down the comms release team when iterating in the notes,
adding missing labels, etc. It also makes development of the script
painful when testing it manually.

The default is 10 concurrent routines, but this can be changed with
command flag.

The default value of 10 doesn't seem to run into rate limiting problems.
I had to bump it to a 100 and run the command multiple times in a row to
be able to get rate limited.