Kustomize: Update deprecated syntax
This commit updates the following:

- patchesStrategicMerge -> patches
- patchesJson6902 -> patches
- vars and varReference -> replacements
- bases -> resources

Most of this is straight forward, but the vars -> replacements change is
a bit complicated. I have taken inspiration from kubebuilder for how to
do the change. In particular I changed the name of the secret that holds
the certificate to be static. Previously it was set partially from a
variable. I believe it would be unnecessarily complicated to keep this
behavior and that a static name does not take away much.