Watch external objects for machine before deleting
This fixes a race condition in the machine controller when the
controller is restarted after a machine has been marked for deletion but
before the infra machine or bootstrap config are deleted. In that case,
the controller doesn't have watches on those external objects, so the
reconciliation is not triggered once they are deleted. This means that
Machines stay in Deleting state for the resync period, the default is 10
minutes.