Update reference to clusterctl config file
The ProxMox initialization steps mention adding the IPAM provider to the
config file. This is the first mention of the config file in the doc,
with very little context as to what it is or how to work with it.

This updates the text to also provide a link to the clusterctl config
documentation to make it easier to find out more about what is being
asked. It also updates the path given to the current XDG-based path
instead of the older `$HOME/.cluster-api` directory path.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>