clusterctl: Add 0 default to worker-machine-count help
Cobra nicely adds the default value for arguments, except in the case
where the default is 0 or empty. This causes the help output for
`clusterctl generate cluster` `--worker-machine-count` to not include
its default of 0, unlike `--control-plane-machine-count` that has a
default of 1.

This change explicitly adds the default value to the help text for
`--worker-machine-count` to make it easier to discover what to expect
when running the command without this argument.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>