fix: verify that there is a release for the tag
Now we attempt to retrieve provider release data first via the
go-proxy client. In case of failure, we would resort to direct polling
from GitHub. However, the go-proxy client only lists existing git
tags and does not return detailed information about github releases.
This behavior leads to a failure in clustercl if there's no release
for the latest tag.

To address this, we try to get a release for the tag first, and return
the highest version with a corresponding release.

For the sake of performance we do only 5 attempts to find a release,
and return an error after that.