internal/webhooks: make MDT.replicas and autoscaler mut. exclusive
Add new validation functions in cluster.go
and clusterclass.go:
- ValidateAutoscalerAnnotationsForCluster
- ValidateAutoscalerAnnotationsForClusterClass

..ForCluster is called from validateTopology().
It makes sure that a given Topology does not contain
MachineDeploymentToplogy objects under Workers, that have
both a "replicas" field set to non-nil and an autoscaler
min/max annotation present.

Also optionally it checks if there are annotations that are set
on the ClusterClass's MachineDeploymentClass for this
MachineDeploymentTopology.

..ForClusterClass accepts a list of Clusters that use
a certain ClusterClass. Inline, it uses the ..ForCluster
function and the same validation as above is performed.
It is called from validate() for a CC on update.

Additionally:
- Add a new builder utility function
  MachineDeploymentTopologyBuilder#WithAnnotations().
- Add 100% test coverage for the new validation functions.