Flag for old infra machine naming
There is a breaking change to the way InfraMachines are named in
v1.7.0. This commit adds a flag to the KCP and core controller for
getting the old behavior back. It defaults to disabled so the new
behavior stays. It is also marked as deprecated and to be removed in v1.9.

Signed-off-by: Lennart Jern <lennart.jern@est.tech>