test: Ensure ownerRef assertions for all Kinds are evaluated
The assertFuncs are passed as a list. Previously, if two assertFuncs
were defined for the same Kind, the assertFunc that appeared later
in the list would overwrite the one that appeared earlier.

If someone introduced a second assertFunc for a given Kind, they would
receive no signal that the first assertFunc was overwritten. That would
result in missed assertions.

With this change, if two or more assertFuncs are defined for the same
Kind, both are evaluated.