fix: Correctly handle concurrent updates to ClusterResourceSetBinding
The existing code does not do optimistic locking on the CRSBinding via
`resourceVersion` and hence concurent updates (patches) overwrite each
other, leading to races and inconsistent state of the CRSBinding.

This commit fixes that by forcing optimistic locking via the
controller-runtime client patch options. The downside to this is that
it leads to more reconciles and log output of failed updates, but
this is a much better situation than having inconsistent and inaccurate
state stored in CRSBinding.