Deprecate obsolete errors pkg
The /errors package has its origin in when capi providers were machineActuators that needed to vendor core capi to function.
There's no usage recommendations and value is questionable since we moved to CRDs and conditions for interoperability between core and providers.
I think we should deprecate it and if there's any use case relying on it we should support it via conditions