Set Machine's BootstrapReady when there is no ConfigRef
If there is no ConfigRef but the bootstrap data secret is set by the
user (instead of a bootstrap provider), then BootstrapReady should be
true. This is the case for MachineSet, and was originally the case for
Machine since 5113f801f130f2d8f0a1b90403b7d868c696c1b8. However, in
d93eadcbc06ad6ef5131e2a6df1673be4397e525 this changed as a side effect
of ensuring that bootstrap config object can continue to be reconciled
after the bootstrap provider has produced the bootstrap data secret.

This change ensures that, once a bootstrap data secret exists, in the
case of a ConfigRef it can still be reconciled, while in the case there
is no ConfigRef, BootstrapReady is set.

Signed-off-by: Zane Bitter <zbitter@redhat.com>