🐛 Align random deletion policy behavior with the rest regarding deletion annotations
Right now, the `oldest` and `newest` deletion policies prioritize machines that have the `cluster.x-k8s.io/delete-machine` annotation over unhealthy machines when deciding which machines need to be deleted. This wasn't the case with the `random` deletion policy. This change aligns the latter to the rest, making it honor the deletion annotation.