[release-1.7] 🌱 clusterctl: always run crd migration if possible to reduce conversion webhook usage (#11513)
* clusterctl: always run crd migration if possible to reduce conversion webhook usage

If objects are still stored in the old version, every time an object is requested
the kube-apiserver will call the conversion webhook. Instead of lots of conversion
webhook calls we once do the migration if possible.

Also increases the timeout used for the migrations: Before running CRD migration
clusterctl usually upgrades cert-manager which may lead to updating Certificates and
rolling out new certificates to our controllers. There is a chance that this can take
up to 90s in which conversion, validating or mutatingwebhooks may be unavailable.
Because the migration gets run more aggresively during upgrades this also increases
the relevant timeouts.

* review fixes

* review fixes

---------

Co-authored-by: Christian Schlotter <christian.schlotter@broadcom.com>