Fix multiline Ready condition in clusterctl describe
The multiline feature in clusterctl has a bug in the multiline conditions.
If the Ready condition message has multipile lines the multiline prefix is not calculated properly.
To fix this we can use the same logic as for the other conditions and verify if the object is the last object in the tree.

Signed-off-by: Tobias Giese <tgiese@nvidia.com>