feat: set new replica fields for machine pools
This updates the MachinePool reconciliation to set the values for the
replica fields in the status section to be consistent with things
like the MachineSets.

If a infra machine pool doesn't have "machine pool machines" then we
can't realistically set the values for all the fields, so in this case
we set the values the same as the replicas.

This change also introduces setting the "UptoDate" condition on the
Machines for the first time. It currently only takes the deletion
timestamp into account when determining if a machine is outdated. We
can iterate on this going forward.

Signed-off-by: Richard Case <richard.case@outlook.com>