feat: add check metadata version CLI utility
This adds a CLI utility which given a version number will
check that there is a release series defined in a metadata file for
that version.

The main use case for this is a CI check. When building a release of a
provider that is for a new major or minor version its required that a
new release series is added to the providers metadata file. This ensures
that the correct version of the provider is installed when initializing
a management cluster.

A common failure for release pipelines is that the metadata file hasn't
been updated.

Signed-off-by: Richard Case <richard.case@outlook.com>