Verify providers need upgrade before applying
If a user runs `upgrade plan` and the results show there are no
available upgrades, the expectation is they then do not attempt to apply
the plan. But nothing prevents them from doing so.

If a non-upgrade plan is applied, the current apply logic does not
handle the case where the "next version" is empty. To make this more
robust, and to protect against users trying to apply plans with no
available upgrades, this performs a quick validation on the plan
providers to make sure they have a version to upgrade to. If not, the
`upgrade apply` call ends up being a no-op and avoids potentially
confusing error output.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>